separate historic artist schema
